requirements.txt
annotated-types==0.7.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
dotenv==0.9.9
fonttools==4.60.1
gitdb==4.0.12
GitPython==3.1.45
google-cloud==0.34.0
idna==3.11
joblib==1.5.2
kiwisolver==1.4.9
matplotlib==3.10.7
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
platformdirs==4.5.0
protobuf==6.33.1
pydantic==2.12.4
pydantic_core==2.41.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
sentry-sdk==2.45.0
six==1.17.0
smmap==5.0.2
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wandb==0.23.0
xgboost==3.1.2
google-cloud-storage==2.18.2